requirements.txt
easydict==1.6
joblib==0.10.3
numpy==1.12.0
Pillow==4.0.0
tensorflow-gpu==1.10.1
# ROS
PyYAML
rospkg
catkin_pkg
